requirements.txt
python-dotenv>=1.0.1
google-api-python-client>=2.130.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
openai>=1.35.0
